authorization-server/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
pyjwt
cryptography